requirements.txt
alembic==1.7.5
anyio==3.4.0
apache-airflow==2.2.2
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
argon2-cffi==21.1.0
attrs==20.3.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
constantly==15.1.0
croniter==1.0.15
cryptography==36.0.0
cssselect==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
Flask==1.1.4
Flask-AppBuilder==3.4.0
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.28.2
graphviz==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
h2==3.2.0
heroku==0.1.4
hpack==3.0.0
httpcore==0.14.3
httpx==0.21.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
incremental==21.3.0
inflection==0.5.1
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
iso8601==1.0.2
isodate==0.6.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==1.1.0
jedi==0.18.1
Jinja2
jmespath==0.10.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.12.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lockfile==0.12.2
lxml==4.6.4
Mako==1.1.6
Markdown==3.3.6
MarkupSafe
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.21.4
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
priority==1.3.0
prison==0.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
Protego==0.1.16
psutil==5.8.0
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.10.0
PyJWT==1.7.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
queuelib==1.6.2
requests==2.26.0
rfc3986==1.5.0
rich==10.14.0
scipy==1.7.3
Scrapy==2.5.1
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==21.1.0
setproctitle==1.2.2
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
spotipy==2.19.0
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
Twisted==21.7.0
typing-extensions==4.0.0
unicodecsv==0.14.1
urllib3==1.26.7
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
zope.interface==5.4.0
